move txn.Mounter to entry
